updates to allow signing of txs generated by counterparty-core 10.9+
As of counterparty-core 10.9, the core developers have determined that passing `script_pubkey` as has been done for the past 10+ years is now considered a bug. This resulted in signing libraries being unable to sign transactions generated with counterparty-core 10.9+ (electrum, bitcore, sparrow, bitcoinjs, etc)

In addition, in counterparty-core 10.9+ the transaction version has been forced from 1 to 2, which also caused problems for above mentioned signing libraries. Version 2 is typically only used when time locks are required, which is not the case in counterparty transactions. Since counterparty will need to use time locks in atomic swaps in the future, the cp core devs felt it was most appropriate to force ALL transactions to use version 2, instead of just those that use time locks, which in turn caused additional signing issues

This update resolves these issues by:
- Requesting tx data for every input in a tx and updating the tx with the input type
- Converting tx back to version 1 so that libraries are able to sign them again

TODO:
At some point in the future, if I ever decide to support atomic swaps in FreeWallet, will need to update this signing code to actually sign version 2 txs, but considering how unstable atomic swaps are and that there are a number of users who have lost funds and assets by spending utxos with assets bound to them, this is probably not something that will be included in FreeWallet anytime soon.